Fix redraw issue.
I was annoyed by redrawing issues that happened when switch buffer,
e.g. when using tag command or switching between headers and sources.
I was fixing this by pressing CTRL-L for a while.

But today I removed all my plugins to identify the root of the evil.
Then adding the half of them, another half of the half, ...
actually a kind of binary search ;-)

Unfortunately I figured out, that my own plugin vim-konsole was causing
this issue. Sorry for that.
Now this issues should have been fixed.